[tests] splitted tests
