+++ Podfile
@@ -6,6 +6,8 @@ target 'Akane' do
   podspec
 
   target 'AkaneTests' do
+    inherit! :search_paths
+
     pod "Quick"
     pod "Nimble", '~> 5.x'
   end
@@ -13,4 +15,11 @@ end
 
 target 'AkaneBond' do
   podspec
+
+  target 'AkaneBondTests' do
+    inherit! :search_paths
+
+    pod "Quick"
+    pod "Nimble", '~> 5.x'
+  end
 end
